requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.6.0
django-extensions==3.1.0
django-import-export==2.5.0
django-mysql==3.10.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
et-xmlfile==1.0.1
graphviz==0.16
idna==2.10
isort==5.7.0
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
pycparser==2.20
pydot==1.4.1
pydot-ng==2.0.0
PyJWT==2.0.0
pylint==2.6.0
pylint-django==2.4.2
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0